Finished the calculator
Calculator is relatively bug-free. Removed the operation class from
the equals sign in index.html and designated a yellow background similar
to the operation style keys in style.css to account for the loss of
the operation class css. Created restrictions to the equals key that
will prevent any negative interactions that may have occurred from
pressing "=" prematurely.